+++ node/resolvers/items.ts
@@ -1,9 +1,7 @@
-import { map } from 'bluebird'
-import { Logger } from '@vtex/api'
-
 import { SearchGraphQL } from '../clients/searchGraphQL'
 import { fixImageUrl } from '../utils/image'
 import { addOptionsForItems } from '../utils/attachmentsHelpers'
+import { OrderFormIdArgs } from '../utils/args'
 
 const GOCOMMERCE = 'gocommerce'
 
@@ -40,59 +38,58 @@ const getVariations = (skuId: string, skuList: any[]) => {
   }))
 }
 
-export const adjustItems = ({
-  platform,
-  items,
-  searchGraphQL,
-  logger,
-}: {
-  platform: string
-  items: OrderFormItem[]
-  searchGraphQL: SearchGraphQL
-  logger: Logger
-}) =>
-  map(items, async (item: OrderFormItem) => {
-    let product = null
+export const root = {
+  Item: {
+    imageUrls: (item: OrderFormItem, _: unknown, ctx: Context) => {
+      return fixImageUrl(item.imageUrl, ctx.vtex.platform)
+    },
+    skuSpecifications: async (
+      item: OrderFormItem,
+      _: unknown,
+      ctx: Context
+    ) => {
+      const {
+        vtex: { platform, logger },
+        clients: { searchGraphQL },
+      } = ctx
 
-    try {
-      product = await getProductInfo(platform, item, searchGraphQL)
-    } catch (err) {
-      logger.warn({
-        message: 'Error when communicating with vtex.search-graphql',
-        error: err,
-      })
-    }
+      let product = null
 
-    return {
-      ...item,
-      imageUrls: fixImageUrl(item.imageUrl, platform),
-      name: product?.productName ?? item.name,
-      skuSpecifications: getVariations(item.id, product?.items ?? []),
-    }
-  })
+      try {
+        product = await getProductInfo(platform, item, searchGraphQL)
+      } catch (err) {
+        logger.warn({
+          message: 'Error when communicating with vtex.search-graphql',
+          error: err,
+        })
+      }
+
+      return getVariations(item.id, product?.items ?? [])
+    },
+  },
+}
 
 export const mutations = {
   addToCart: async (
     _: unknown,
-    {
-      items,
-      marketingData = {},
-    }: {
+    args: {
       items: OrderFormItemInput[]
       marketingData: Partial<OrderFormMarketingData>
-    },
+    } & OrderFormIdArgs,
     ctx: Context
   ): Promise<CheckoutOrderForm> => {
     const {
       clients,
-      vtex: { orderFormId, logger },
+      vtex,
+      vtex: { logger },
     } = ctx
+    const { orderFormId = vtex.orderFormId, items, marketingData = {} } = args
 
     const { checkout } = clients
     const shouldUpdateMarketingData =
       Object.keys(marketingData ?? {}).length > 0
 
-    const { items: previousItems } = await checkout.orderForm()
+    const { items: previousItems } = await checkout.orderForm(orderFormId!)
     const cleanItems = items.map(({ options, ...rest }) => rest)
     const withOptions = items.filter(
       ({ options }) => !!options && options.length > 0
@@ -132,7 +129,7 @@ export const mutations = {
         previousItems
       )
 
-      return checkout.orderForm()
+      return checkout.orderForm(orderFormId!)
     }
 
     return newOrderForm
@@ -140,17 +137,18 @@ export const mutations = {
 
   updateItems: async (
     _: unknown,
-    { orderItems }: { orderItems: OrderFormItemInput[] },
+    args: {
+      orderItems: OrderFormItemInput[]
+      splitItem: boolean
+    } & OrderFormIdArgs,
     ctx: Context
   ): Promise<CheckoutOrderForm> => {
-    const {
-      clients,
-      vtex: { orderFormId },
-    } = ctx
+    const { clients, vtex } = ctx
+    const { orderFormId = vtex.orderFormId, orderItems, splitItem } = args
     const { checkout } = clients
 
     if (orderItems.some((item: OrderFormItemInput) => !item.index)) {
-      const orderForm = await checkout.orderForm()
+      const orderForm = await checkout.orderForm(orderFormId!)
 
       const idToIndex = orderForm.items.reduce(
         (acc: Record<string, number>, item: OrderFormItem, index: number) => {
@@ -171,88 +169,104 @@ export const mutations = {
 
     const newOrderForm = await clients.checkout.updateItems(
       orderFormId!,
-      orderItems
+      orderItems,
+      splitItem
     )
 
     return newOrderForm
   },
   addItemOffering: async (
     _: unknown,
-    { offeringInput }: { offeringInput: OfferingInput },
+    args: { offeringInput: OfferingInput } & OrderFormIdArgs,
     ctx: Context
   ): Promise<CheckoutOrderForm> => {
-    const {
-      clients,
-      vtex: { orderFormId },
-    } = ctx
+    const { clients, vtex } = ctx
+    const { offeringInput, orderFormId = vtex.orderFormId } = args
 
     const newOrderForm = await clients.checkout.addItemOffering(
       orderFormId!,
       offeringInput.itemIndex,
       offeringInput.offeringId,
-      offeringInput.offeringInfo,
+      offeringInput.offeringInfo
     )
 
     return newOrderForm
   },
   removeItemOffering: async (
     _: unknown,
-    { offeringInput }: { offeringInput: OfferingInput },
+    args: { offeringInput: OfferingInput } & OrderFormIdArgs,
     ctx: Context
   ): Promise<CheckoutOrderForm> => {
-    const {
-      clients,
-      vtex: { orderFormId },
-    } = ctx
+    const { clients, vtex } = ctx
+    const { offeringInput, orderFormId = vtex.orderFormId } = args
 
     const newOrderForm = await clients.checkout.removeItemOffering(
       orderFormId!,
       offeringInput.itemIndex,
-      offeringInput.offeringId,
+      offeringInput.offeringId
     )
 
     return newOrderForm
   },
   addBundleItemAttachment: async (
     _: unknown,
-    { bundleItemAttachmentInput }: { bundleItemAttachmentInput: BundleItemAttachmentInput },
+    args: {
+      bundleItemAttachmentInput: BundleItemAttachmentInput
+    } & OrderFormIdArgs,
     ctx: Context
   ): Promise<CheckoutOrderForm> => {
-    const {
-      clients,
-      vtex: { orderFormId },
-    } = ctx
+    const { clients, vtex } = ctx
+    const { bundleItemAttachmentInput, orderFormId = vtex.orderFormId } = args
 
     const newOrderForm = await clients.checkout.addBundleItemAttachment(
       orderFormId!,
       bundleItemAttachmentInput.itemIndex,
       bundleItemAttachmentInput.bundleItemId,
       bundleItemAttachmentInput.attachmentName,
-      bundleItemAttachmentInput.attachmentContent,
+      bundleItemAttachmentInput.attachmentContent
     )
 
     return newOrderForm
   },
   removeBundleItemAttachment: async (
     _: unknown,
-    { bundleItemAttachmentInput }: { bundleItemAttachmentInput: BundleItemAttachmentInput },
+    args: {
+      bundleItemAttachmentInput: BundleItemAttachmentInput
+    } & OrderFormIdArgs,
     ctx: Context
   ) => {
-    const {
-      clients,
-      vtex: { orderFormId },
-    } = ctx
+    const { clients, vtex } = ctx
+    const { bundleItemAttachmentInput, orderFormId = vtex.orderFormId } = args
 
     const { data } = await clients.checkout.removeBundleItemAttachment(
       orderFormId!,
       bundleItemAttachmentInput.itemIndex,
       bundleItemAttachmentInput.bundleItemId,
       bundleItemAttachmentInput.attachmentName,
-      bundleItemAttachmentInput.attachmentContent,
+      bundleItemAttachmentInput.attachmentContent
     )
 
     return data
-  }
+  },
+  setManualPrice: async (
+    _: unknown,
+    args: { input: { itemIndex: number; price: number } } & OrderFormIdArgs,
+    ctx: Context
+  ): Promise<CheckoutOrderForm> => {
+    const { clients, vtex } = ctx
+    const {
+      input: { itemIndex, price },
+      orderFormId = vtex.orderFormId,
+    } = args
+
+    const newOrderForm = await clients.checkoutAdmin.setManualPrice(
+      orderFormId!,
+      itemIndex,
+      price
+    )
+
+    return newOrderForm
+  },
 }
 
 interface OfferingInput {