+++ node/utils/shipping.ts
@@ -1,5 +1,4 @@
-import flatten from 'lodash/flatten'
-import uniq from 'lodash/uniq'
+// eslint-disable-next-line no-restricted-imports
 import uniqBy from 'lodash/uniqBy'
 
 import {
@@ -12,6 +11,7 @@ import {
   hasDeliveryOption,
 } from './delivery-options'
 import { Clients } from '../clients'
+import { DELIVERY, PICKUP_IN_POINT } from '../constants'
 
 export const getShippingData = (
   address: CheckoutAddress,
@@ -34,28 +34,40 @@ export const getShippingData = (
   return requestPayload
 }
 
-export const selectDeliveryOption = ({
+export const selectShippingOption = ({
   shippingData,
-  deliveryOptionId,
+  slaId,
+  itemId,
+  deliveryChannel,
 }: {
-  shippingData: ShippingData
-  deliveryOptionId: string
+  shippingData: ShippingData | null
+  slaId: string
+  itemId?: string
+  deliveryChannel: string
 }) => {
-  const logisticsInfoWithSelectedDeliveryOption = shippingData.logisticsInfo.map(
-    (li: LogisticsInfo) => ({
-      ...li,
-      selectedSla: hasDeliveryOption(li.slas, deliveryOptionId)
-        ? deliveryOptionId
-        : li.selectedSla,
-    })
-  )
+  const logisticsInfoWithSelectedDeliveryOption =
+    shippingData?.logisticsInfo.map((li: LogisticsInfo) => {
+      return {
+        ...li,
+        selectedDeliveryChannel: deliveryChannel,
+        selectedSla:
+          hasDeliveryOption(li.slas, slaId) &&
+          (itemId != null ? li.itemId === itemId : true)
+            ? slaId
+            : li.selectedSla,
+      }
+    }) ?? []
 
   const deliveryAddress = getSelectedDeliveryAddress(
-    shippingData.selectedAddresses
+    shippingData?.selectedAddresses ?? []
   )
 
+  if (!deliveryAddress) {
+    return shippingData
+  }
+
   return getShippingData(
-    deliveryAddress!,
+    deliveryAddress,
     logisticsInfoWithSelectedDeliveryOption
   )
 }
@@ -64,9 +76,13 @@ export const selectAddress = ({
   shippingData,
   address,
 }: {
-  shippingData: ShippingData
+  shippingData: ShippingData | null
   address: CheckoutAddress
-}): ShippingData => {
+}): ShippingData | null => {
+  if (!shippingData) {
+    return null
+  }
+
   return {
     ...shippingData,
     selectedAddresses: [address],
@@ -78,35 +94,34 @@ export const getShippingInfo = async ({
   orderForm,
 }: {
   clients: Clients
-  orderForm: CheckoutOrderForm
+  orderForm: Pick<
+    CheckoutOrderForm,
+    'shippingData' | 'totalizers' | 'orderFormId' | 'value'
+  >
 }) => {
-  const logisticsInfo =
-    orderForm.shippingData && orderForm.shippingData.logisticsInfo
+  const logisticsInfo = orderForm.shippingData?.logisticsInfo ?? []
 
-  const countries = uniq(
-    flatten(logisticsInfo ? logisticsInfo.map(item => item.shipsTo) : [])
+  const countries = Array.from(
+    new Set(logisticsInfo.flatMap(item => item.shipsTo)).values()
   )
 
-  const availableAddresses =
-    (orderForm.shippingData && orderForm.shippingData.availableAddresses) || []
+  const availableAddresses = orderForm.shippingData?.availableAddresses ?? []
 
   const selectedAddress =
     orderForm.shippingData &&
-    getSelectedDeliveryAddress(orderForm.shippingData.selectedAddresses)!
+    getSelectedDeliveryAddress(orderForm.shippingData.selectedAddresses)
 
   const availableItemsLogisticsInfo = logisticsInfo
-    ? logisticsInfo.filter((item: LogisticsInfo) => item.slas.length)
+    ? logisticsInfo.filter((item: LogisticsInfo) => item.slas.length > 0)
     : []
 
   const deliveryOptions = uniqBy(
-    flatten(
-      availableItemsLogisticsInfo.map((item: LogisticsInfo) => item.slas)
-    ),
+    availableItemsLogisticsInfo.flatMap((item: LogisticsInfo) => item.slas),
     'id'
   )
 
-  // Since at this time Shipping does not show Pickup Points or
-  // Scheduled Delivery/Pickup Options we are filtering from results.
+  // Since at this time shipping does not show scheduled delivery/pickup
+  // we are filtering from them results.
   // Also we will filter deliveryOptions which does not apply to all LogisticsInfo.
   const filteredDeliveryOptions = filterDeliveryOptions(
     deliveryOptions,
@@ -119,7 +134,7 @@ export const getShippingInfo = async ({
   )
 
   const selectedDeliveryOption = updatedDeliveryOptions.find(
-    option => option.isSelected
+    option => option.isSelected && option.deliveryChannel === DELIVERY
   )
 
   const shippingTotalizer = orderForm.totalizers.find(
@@ -131,8 +146,9 @@ export const getShippingInfo = async ({
     shippingTotalizer &&
     selectedDeliveryOption.price !== shippingTotalizer.value
   ) {
-    const newShippingData = selectDeliveryOption({
-      deliveryOptionId: selectedDeliveryOption.id,
+    const newShippingData = selectShippingOption({
+      slaId: selectedDeliveryOption.id,
+      deliveryChannel: DELIVERY,
       shippingData: orderForm.shippingData,
     })
 
@@ -149,7 +165,23 @@ export const getShippingInfo = async ({
   return {
     availableAddresses,
     countries,
-    deliveryOptions: updatedDeliveryOptions,
+    deliveryOptions: updatedDeliveryOptions.filter(
+      ({ deliveryChannel }) => deliveryChannel === DELIVERY
+    ),
+    pickupOptions: updatedDeliveryOptions
+      .filter(({ deliveryChannel }) => deliveryChannel === PICKUP_IN_POINT)
+      .map(pickupOption => {
+        return {
+          id: pickupOption.id,
+          price: pickupOption.price,
+          estimate: pickupOption.estimate,
+          isSelected: pickupOption.isSelected,
+          friendlyName: pickupOption.sla.pickupStoreInfo.friendlyName,
+          additionalInfo: pickupOption.sla.pickupStoreInfo.additionalInfo,
+          storeDistance: pickupOption.sla.pickupDistance,
+          transitTime: pickupOption.sla.transitTime,
+        }
+      }),
     selectedAddress,
   }
 }